Ran tuning for RF, HPC problems understood
I think the problem with running code on the HPC is the NUMA effects, since increasing the number of cores also increases the number of nodes needed, and so it ends up taking more time with more cores, as confirmed by the benchmark results. Notice in the code here, they I don't use ray backend since I attemtped this and it ended up taking more time, which may be due to ray then attempting to do it's own backend, even though sklearn already seems to have this implemented in some way. We probably won't be using the HPC, since I ran the hyperparameter tuning on my own computer using 4 cores, and I got the results after about 11.2 hours, though this isn't written to the file as I accidentally closed the chrome tab, so I had to reload the predictions csv file (not uploaded here) to run the last few cells.