Got cons preds, adjusted fonts of figs
I ran the same hyperparameter tuning code for the consumption data, which took only around 10 hours, and seems to have much lower mse, which makes sense since there was very little variation in the consumption pattern. I also adjusted the font size for some the graphs in the Filling missing vals and in the Statistical analysis, simply because they were hard to read on the overleaf document.