DP added
Once again just using code from the Otovo project. Unsure if the spot price data is in øre, because the prices look insane lol. We'll also need to take a closer look at the DP model to make sure it's doing the right thing, because although it looks correct, I'd just like to be entirely sure.