Finished V2, started V3
In BatterySimV2, I attempted to clean all the code from the Otovo project, but as this proved cumbersome to do for the battery class, I decided to make a V3, where the battery charge method is comepletely overhauled and now easier to read. I will also completely do-over the DPModel: The new model shouldn't buy and sell as an action, but rather just charge/discharge the battery.

Here, I have finished the battery, made some changes to the logics, and have made the DPModel ready for remaking.