Functions readability, priceopt cont.
I made the Actions space of the DPModel more readable, as well as removing unused parameters in get_price. I also changed logic_rollout to take actions, so that the actions from DP aren't contained in its battery (as that was rather silly).

I implemented these changes in BatterySimV3 as well as Price-Optimizaiton, where I also implemented using the predictions properly, but have still yet to fully implement the desired system (but I'm working on it).