Implemented emissions
I made it a part of the merge function. I Also made a second DP model that has emissions as cost function. I made emissions a part of the logic rollouts too. I also managed to run it for a week in battery sim (no predictions), though it stopped running on my laptop using predictions. Will run it again now on desktop PC