Added switch to hardcode low prices
Line is marked by a comment. This push makes sure that the prices are high and it is then up to the individual to find the line in this file and uncomment it if low prices are wanted. Also added file "high_to_low.ipynb" where the difference can be seen.